more stuff

+++ ex08/powerset.cpp
@@ -1,17 +1,16 @@
-#include <cstdint>
-#include <vector>
+#include "powerset.hpp"
 
 std::vector<std::vector<int32_t> >	powerset(std::vector<int32_t>	set)
 {
 	std::vector<std::vector<int32_t> >	powset;
 	int									n = 1;
-	for (int i = 0; i < set.size(); i++)
+	for (size_t i = 0; i < set.size(); i++)
 		n *= 2;
 	for (int i = 0; i < n; i++)
 	{
 		std::vector<int32_t>	pset;
 		int						x = i;
-		for (int j = 0; j < set.size(); j++)
+		for (size_t j = 0; j < set.size(); j++)
 		{
 			if (x % 2)
 				pset.push_back(set[j]);
